+++ app/Http/Controllers/TenantController.php
@@ -44,6 +44,9 @@ use App\facilityModel as facilityModel;
 	public function rentRoomVehicles(Request $request){
 		$user = $request->session()->get('user');
 
+		if (is_null($user)){
+			return redirect()->action('MainController@index');
+		}
 		//Gets building categories
 		$buildingCategories = buildingCategory::all();
 	//	$buildings = buildingModel::where('landlordID', '=', $user[0]->ID)->get();
@@ -78,6 +81,9 @@ use App\facilityModel as facilityModel;
 	public function myBookings(Request $request){
 		$user = $request->session()->get('user');
 
+		if (is_null($user)){
+			return redirect()->action('MainController@index');
+		}
 		$bookings = bookingModel::where('tenantID', '=', $user[0]->id)->orderBy('created_at', 'DESC')->get();
 
 		return view('pages.mybookings',  array('user' => $user, 'bookings' => $bookings));
@@ -92,6 +98,10 @@ use App\facilityModel as facilityModel;
 
 
 		$user = $request->session()->get('user');
+
+		if (is_null($user)){
+			return redirect()->action('MainController@index');
+		}
 		return view('pages.tenantGetDirections',  array('user' => $user));
 
 		
@@ -105,6 +115,9 @@ use App\facilityModel as facilityModel;
 	$error=true;
 	$user = $request->session()->get('user');
 	
+		if (is_null($user)){
+			return redirect()->action('MainController@index');
+		}
 
 
 	if(isset($request['id'])){
@@ -219,7 +232,10 @@ use App\facilityModel as facilityModel;
 	}
 
 	public function deleteBooking(Request $request){
-
+		$user = $request->session()->get('user');
+		if (is_null($user)){
+			return false;
+		}
 		if ($request['id'] != null){
 			$deleteBookings = bookingModel::where('id', '=', $request['id'])->delete();
